Rearragned EDA and Added OHE
Now EDA for classification is brought up, still needs some touching up in terms of individual variable exploring.

Fpr OHE, I have added the rest of the numeric variables and the categorical variables, should use response_ohe_up as the most updated resampled, encoded dataframe, with the response and predictor variables in it.